requirements.txt
asgiref==3.6.0
cloudinary==1.25.0
dj-database-url==0.5.0
Django==3.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
env.py==0.2.2
gunicorn==20.1.0
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2022.7.1
sqlparse==0.4.1
timedelta==2020.12.3
whitenoise==6.4.0